requirements.txt
boto3==1.16.59
botocore==1.19.59
et-xmlfile==1.0.1
jdcal==1.4.1
jmespath==0.10.0
openpyxl==3.0.6
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.2
xlrd==2.0.1
